requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.0.1
bert-score==0.3.13
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decorator==4.4.2
decord==0.6.0
deepspeed==0.16.4
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flash_attn==2.7.4.post1
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio==3.50.2
gradio_client==0.6.1
groovy==0.1.2
grpcio==1.73.1
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.4.5
ipaddress==1.0.23
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
liger_kernel==0.5.10
-e git+https://github.com/Eureka-Maggie/ROMA.git@97e0a4cb5c1a7c372a7d6470ab8a9266d9163505#egg=llamafactory
llvmlite==0.44.0
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
mutagen==1.47.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.1
omegaconf==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
peft==0.15.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
polars==1.31.0
pooch==1.8.2
portalocker==3.2.0
proglog==0.1.12
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
pyzmq==27.0.0
qwen-omni-utils==0.0.4
qwen-vl-utils==0.0.11
referencing==0.37.0
regex==2024.11.6
requests==2.28.2
rfc3986==1.5.0
rich==14.0.0
rouge-chinese==1.0.3
rpds-py==0.30.0
ruff==0.11.8
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.1
tqdm==4.65.2
traitlets==5.14.3
transformers @ git+https://github.com/Eureka-Maggie/transformers.git@roma_patch
triton==3.2.0
trl==0.9.6
typer==0.15.3
typing_extensions==4.13.2
tyro==0.8.14
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
volcengine-python-sdk==4.0.4
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.17.2
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
